+++ Anarchy/UI/types/ui.d.ts
@@ -194,6 +194,7 @@ declare module "cs2/ui" {
   	"Change Time Scale": Action1D;
   	"Switch Page": Action1D;
   	"Default Tool": Action;
+  	"Default Tool UI": Action;
   	"Tool Options": Action;
   	"Switch Toolmode": Action;
   	"Toggle Snapping": Action;